versi 8.1

+++ system/proses_edit_pengajuan.php
@@ -7,10 +7,18 @@ $jenis_pengajuan = $_POST['jenis_pengajuan'];
 $biaya = $_POST['biaya'];
 $alasan = $_POST['alasan'];
 $keterangan = $_POST['keterangan'];
+$ubah_foto = $_POST['ubah_foto'];
 
 $tanggal= mktime(date("m"),date("d"),date("Y"));
 $tgl = date("Y-m-d", $tanggal);
-if(isset($_POST['ubah_foto'])){ 
+
+
+		$query = "SELECT * FROM pengajuan WHERE id_pengajuan='".$id."'";
+		$sql = mysqli_query($con, $query); 
+		$data = mysqli_fetch_array($sql); 
+
+
+if(isset($_POST['tambah_foto'])){
 	$foto = $_FILES['foto']['name'];
 	$tmp = $_FILES['foto']['tmp_name'];
 
@@ -20,9 +28,31 @@ if(isset($_POST['ubah_foto'])){
 
 	if(move_uploaded_file($tmp, $path)){ 
 
-		$query = "SELECT * FROM pengajuan WHERE id_pengajuan='".$id."'";
+		$query = "UPDATE pengajuan SET pengajuan='$pengajuan'
+         , jenis_pengajuan='$jenis_pengajuan', tanggal_pengajuan='$tgl'
+         , gambar='$fotobaru', biaya='$biaya',alasan='$alasan'
+         , keterangan ='$keterangan' WHERE id_pengajuan='".$id."'";
 		$sql = mysqli_query($con, $query); 
-		$data = mysqli_fetch_array($sql); 
+		if($sql){ 
+			header("location: ../detail_pengajuan?id=$id"); 
+		}else{
+			echo "Maaf, Terjadi kesalahan saat mencoba untuk menyimpan data ke database.";
+			echo "<br><a href='../pengajuan'>Kembali Ke Form</a>";
+		}
+	}else{
+		echo "Maaf, Gambar gagal untuk diupload.";
+		echo "<br><a href='../pengajuan'>Kembali Ke Form</a>";
+	}
+}
+else if($ubah_foto == "ubah" ){ 
+	$foto = $_FILES['foto']['name'];
+	$tmp = $_FILES['foto']['tmp_name'];
+
+	$fotobaru = date('dmYHis').$foto;
+	
+	$path = "../image/".$fotobaru;
+
+	if(move_uploaded_file($tmp, $path)){ 
 
 		if(is_file("../image/".$data['gambar'])) 
 			unlink("../image/".$data['gambar']); 
@@ -44,7 +74,28 @@ if(isset($_POST['ubah_foto'])){
 		echo "Maaf, Gambar gagal untuk diupload.";
 		echo "<br><a href='../pengajuan'>Kembali Ke Form</a>";
 	}
-}else{ 
+}
+else if ($ubah_foto == "hapus"){
+	
+if(is_file("../image/".$data['gambar'])) 
+			unlink("../image/".$data['gambar']); 
+
+	$query = "UPDATE pengajuan SET pengajuan='$pengajuan'
+         , jenis_pengajuan='$jenis_pengajuan', tanggal_pengajuan='$tgl'
+         , gambar='', biaya='$biaya',alasan='$alasan', keterangan ='$keterangan' 
+         WHERE id_pengajuan='".$id."'";
+	$sql = mysqli_query($con, $query); 
+
+if($sql){ 
+	header("location:../detail_pengajuan?id=$id"); 
+
+}else{
+
+	echo "Data gagal dihapus. <a href='../pengajuan'>Kembali</a>";
+}
+
+}
+else{ 
 
 	$query = "UPDATE pengajuan SET pengajuan='$pengajuan'
          , jenis_pengajuan='$jenis_pengajuan', tanggal_pengajuan='$tgl'
